+++ autocalibration/src/markerless_camera_calibration_controller.py
@@ -29,20 +29,20 @@ class MarkerlessCameraCalibrationController(CameraCalibrationController):
   def generateCalibration(self, sceneobj, msg):
     """! Generates the camera pose.
     @param   sceneobj   Scene object
-    @param   msg        Payload with camera data from percebro
+    @param   msg        Payload with camera frame data
 
     @return  dict       Dictionary containing publish topic and data to publish
     """
     cur_cam_calib_obj = self.cam_calib_objs[sceneobj.id]
     log.info("Calibration configuration:", cur_cam_calib_obj.config)
-    percebro_cam_data = json.loads(msg)
-    pub_data = cur_cam_calib_obj.localize(percebro_cam_data=percebro_cam_data,
+    cam_frame_data = json.loads(msg)
+    pub_data = cur_cam_calib_obj.localize(cam_frame_data=cam_frame_data,
                                           sceneobj = sceneobj)
     if bool(pub_data):
       if pub_data.get('error') == 'True':
         log.error(pub_data.get('message', 'Weak or insufficient matches'))
       publish_topic = PubSub.formatTopic(PubSub.DATA_AUTOCALIB_CAM_POSE,
-                                         camera_id=percebro_cam_data['id'])
+                                         camera_id=cam_frame_data['id'])
       return {'publish_topic': publish_topic,
               'publish_data': json.dumps(pub_data)}