+++ manager/src/static/js/camcanvas.js
@@ -5,8 +5,7 @@
 
 /**
  * @file camcanvas.js
- * @description This file defines the CamCanvas class, which provides functions for displaying
- * the images recieved from percebro.
+ * @description This file defines the CamCanvas class for display the video frames.
  */
 
 import {