+++ tests/security/malformed_data/baddata_gen.py
@@ -42,7 +42,7 @@ def main():
   rootca="/run/secrets/certs/scenescape-ca.pem"
 
   # Location for generated user/passwd from image
-  auth = "/run/secrets/percebro.auth"
+  auth = "/run/secrets/controller.auth"
 
   bad_data_input = open(args.input[0], 'r')
   bad_data = bad_data_input.readlines()